scrapy_yaml_item_exporter/scrapy_yaml_item_exporter.py
from ruamel.yaml import YAML
from scrapy.exporters import BaseItemExporter


class YAMLItemExporter(BaseItemExporter):

    def __init__(self, file, **kwargs):
        self._configure(kwargs, dont_fail=True)
        self.file = file
        self.yaml = YAML()
        self.yaml.encoding = self.encoding

    def export_item(self, item):
        self.yaml.dump([dict(item)], self.file)
